Refactor: Move server logic into base class
